+++ banyand/metadata/client.go
@@ -33,7 +33,11 @@ import (
 	"github.com/apache/skywalking-banyandb/pkg/run"
 )
 
-const flagEtcdEndpointsName = "etcd-endpoints"
+const (
+	// DefaultNamespace is the default namespace of the metadata stored in etcd.
+	DefaultNamespace      = "banyandb"
+	flagEtcdEndpointsName = "etcd-endpoints"
+)
 
 const flagEtcdUsername = "etcd-username"
 
@@ -58,7 +62,6 @@ type clientService struct {
 	etcdTLSCertFile string
 	etcdTLSKeyFile  string
 	schemaRegistry  schema.Registry
-	alc             *allocator
 	closer          *run.Closer
 	endpoints       []string
 }
@@ -69,7 +72,7 @@ func (s *clientService) SchemaRegistry() schema.Registry {
 
 func (s *clientService) FlagSet() *run.FlagSet {
 	fs := run.NewFlagSet("metadata")
-	fs.StringVar(&s.namespace, "namespace", "banyandb", "The namespace of the metadata stored in etcd")
+	fs.StringVar(&s.namespace, "namespace", DefaultNamespace, "The namespace of the metadata stored in etcd")
 	fs.StringArrayVar(&s.endpoints, flagEtcdEndpointsName, []string{"http://localhost:2379"}, "A comma-delimited list of etcd endpoints")
 	fs.StringVar(&s.etcdUsername, flagEtcdUsername, "", "A username of etcd")
 	fs.StringVar(&s.etcdPassword, flagEtcdPassword, "", "A password of etcd user")
@@ -108,9 +111,8 @@ func (s *clientService) PreRun(ctx context.Context) error {
 		return errors.New("node roles is empty")
 	}
 	nodeRoles := val.([]databasev1.Role)
-	ctxRegister, cancel := context.WithTimeout(ctx, time.Second*5)
-	defer cancel()
-	if err = s.schemaRegistry.RegisterNode(ctxRegister, &databasev1.Node{
+	l := logger.GetLogger(s.Name())
+	nodeInfo := &databasev1.Node{
 		Metadata: &commonv1.Metadata{
 			Name: node.NodeID,
 		},
@@ -118,12 +120,20 @@ func (s *clientService) PreRun(ctx context.Context) error {
 		HttpAddress: node.HTTPAddress,
 		Roles:       nodeRoles,
 		CreatedAt:   timestamppb.Now(),
-	}); err != nil {
+	}
+	for {
+		ctxRegister, cancel := context.WithTimeout(ctx, time.Second*10)
+		err = s.schemaRegistry.RegisterNode(ctxRegister, nodeInfo)
+		cancel()
+		if errors.Is(err, context.DeadlineExceeded) {
+			l.Warn().Strs("etcd-endpoints", s.endpoints).Msg("register node timeout, retrying...")
+			continue
+		}
+		if err == nil {
+			l.Info().Stringer("info", nodeInfo).Msg("register node successfully")
+		}
 		return err
 	}
-	s.alc = newAllocator(s.schemaRegistry, logger.GetLogger(s.Name()).Named("allocator"))
-	s.schemaRegistry.RegisterHandler("shard-allocator", schema.KindGroup|schema.KindNode, s.alc)
-	return nil
 }
 
 func (s *clientService) Serve() run.StopNotify {
@@ -165,10 +175,6 @@ func (s *clientService) TopNAggregationRegistry() schema.TopNAggregation {
 }
 
 func (s *clientService) PropertyRegistry() schema.Property {
-	return s.schemaRegistry
-}
-
-func (s *clientService) ShardRegistry() schema.Shard {
 	return s.schemaRegistry
 }