+++ pkg/cmdsetup/liaison.go
@@ -29,8 +29,10 @@ import (
 	"github.com/apache/skywalking-banyandb/banyand/liaison/http"
 	"github.com/apache/skywalking-banyandb/banyand/metadata"
 	"github.com/apache/skywalking-banyandb/banyand/observability"
+	"github.com/apache/skywalking-banyandb/banyand/queue"
 	"github.com/apache/skywalking-banyandb/banyand/queue/pub"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
+	"github.com/apache/skywalking-banyandb/pkg/node"
 	"github.com/apache/skywalking-banyandb/pkg/run"
 	"github.com/apache/skywalking-banyandb/pkg/version"
 )
@@ -43,11 +45,16 @@ func newLiaisonCmd(runners ...run.Unit) *cobra.Command {
 		l.Fatal().Err(err).Msg("failed to initiate metadata service")
 	}
 	pipeline := pub.New(metaSvc)
-	grpcServer := grpc.NewServer(ctx, pipeline, metaSvc)
+	localPipeline := queue.Local()
+	nodeSel, err := node.NewMaglevSelector()
+	if err != nil {
+		l.Fatal().Err(err).Msg("failed to initiate required node selector")
+	}
+	grpcServer := grpc.NewServer(ctx, pipeline, localPipeline, metaSvc, grpc.NewClusterNodeRegistry(pipeline, nodeSel))
 	profSvc := observability.NewProfService()
 	metricSvc := observability.NewMetricService()
 	httpServer := http.NewServer()
-	dQuery, err := dquery.NewService(metaSvc, pipeline)
+	dQuery, err := dquery.NewService(metaSvc, localPipeline, pipeline)
 	if err != nil {
 		l.Fatal().Err(err).Msg("failed to initiate distributed query service")
 	}
@@ -55,6 +62,7 @@ func newLiaisonCmd(runners ...run.Unit) *cobra.Command {
 	units = append(units, runners...)
 	units = append(units,
 		metaSvc,
+		localPipeline,
 		pipeline,
 		dQuery,
 		grpcServer,